+++ src/winsearch.c
@@ -213,14 +213,17 @@ win_toggle_search(bool show, bool focus)
   printf("ctrl/but %d/%d cell %d/%d font h %d s %d\n", ctrl_height, button_width, cell_height, cell_width, font_height, font_size);
 #endif
 
-  const char * search_bar = cfg.search_bar;
+  const wchar * search_bar = cfg.search_bar;
   int pos_close = -1;
   int pos_prev = -1;
   int pos_next = -1;
   int pos_edit = -1;
   int barpos = margin;
-  while (search_bar && * search_bar)
-    switch (* search_bar ++) {
+  wchar * prev_but = _W("◀");
+  wchar * next_but = _W("▶");
+  wchar * close_but = _W("X");
+  while (search_bar && * search_bar) {
+    switch (* search_bar) {
       when 'x' or 'X':
         place_field(& barpos, button_width, & pos_close);
       when '<':
@@ -229,7 +232,46 @@ win_toggle_search(bool show, bool focus)
         place_field(& barpos, button_width, & pos_next);
       when 's' or 'S':
         place_field(& barpos, edit_width, & pos_edit);
+      when 0x25B2 ... 0x25B5 or 0x25C0 ... 0x25C5:
+        place_field(& barpos, button_width, & pos_prev);
+        * prev_but = * search_bar;
+      when 0x25B6 ... 0x25BF:
+        place_field(& barpos, button_width, & pos_next);
+        * next_but = * search_bar;
+      when 0x2190 ... 0x2193 or 0x21D0 ... 0x21D3 or 0x21E0 ... 0x21E3
+        or 0x21E6 ... 0x21E9 or 0x261C ... 0x261F:
+        if ((* search_bar & 0xF) % 6 > 1) {
+          place_field(& barpos, button_width, & pos_next);
+          * next_but = * search_bar;
+        }
+        else {
+          place_field(& barpos, button_width, & pos_prev);
+          * prev_but = * search_bar;
+        }
+      when 0x21FD or 0x21FE or 0x27F5 or 0x27F6 or 0x261A or 0x261B:
+        if ((* search_bar & 0xF) % 5 % 3) {
+          place_field(& barpos, button_width, & pos_next);
+          * next_but = * search_bar;
+        }
+        else {
+          place_field(& barpos, button_width, & pos_prev);
+          * prev_but = * search_bar;
+        }
+      when 0x3008 ... 0x300B:
+        if (* search_bar & 1) {
+          place_field(& barpos, button_width, & pos_next);
+          * next_but = * search_bar;
+        }
+        else {
+          place_field(& barpos, button_width, & pos_prev);
+          * prev_but = * search_bar;
+        }
+      when 0x2717 or 0x2718 /*or 0x274C or 0x2573*/:
+        place_field(& barpos, button_width, & pos_close);
+        * close_but = * search_bar;
     }
+    search_bar ++;
+  }
   place_field(& barpos, button_width, & pos_close);
   place_field(& barpos, button_width, & pos_prev);
   place_field(& barpos, button_width, & pos_next);
@@ -257,15 +299,15 @@ win_toggle_search(bool show, bool focus)
     search_wnd = CreateWindowExA(0, SEARCHBARCLASS, "", WS_CHILD, 0, 0, 0, 0, wnd, 0, inst, NULL);
 
     //__ label of search bar close button; not actually "localization"
-    search_close_wnd = CreateWindowExW(0, W("BUTTON"), _W("X"), WS_CHILD | WS_VISIBLE,
+    search_close_wnd = CreateWindowExW(0, W("BUTTON"), close_but, WS_CHILD | WS_VISIBLE,
                                      pos_close, margin, button_width, ctrl_height,
                                      search_wnd, NULL, inst, NULL);
     //__ label of search bar prev button; not actually "localization"
-    search_prev_wnd = CreateWindowExW(0, W("BUTTON"), _W("◀"), WS_CHILD | WS_VISIBLE,
+    search_prev_wnd = CreateWindowExW(0, W("BUTTON"), prev_but, WS_CHILD | WS_VISIBLE,
                                      pos_prev, margin, button_width, ctrl_height,
                                      search_wnd, NULL, inst, NULL);
     //__ label of search bar next button; not actually "localization"
-    search_next_wnd = CreateWindowExW(0, W("BUTTON"), _W("▶"), WS_CHILD | WS_VISIBLE,
+    search_next_wnd = CreateWindowExW(0, W("BUTTON"), next_but, WS_CHILD | WS_VISIBLE,
                                      pos_next, margin, button_width, ctrl_height,
                                      search_wnd, NULL, inst, NULL);
     search_edit_wnd = CreateWindowExA(WS_EX_CLIENTEDGE, "EDIT", "", WS_CHILD | WS_VISIBLE | WS_TABSTOP | ES_AUTOHSCROLL,